Move mongo function to ./lib
